Fixing listing of posts in the preview.php module
Now, all posts are listed correctly both in index.php single.php. The
right text (excerpt or content) is shown in each entry of the list and
we're doing a single query to decide which is the selected post and
to list them all (with a defined count of posts to list).